+++ src/components/App/App.js
@@ -4,7 +4,6 @@ import React from 'react';
 import {
   Routes,
   Route,
-  useNavigate,
 } from 'react-router-dom';
 
 import PageWrapper from '../PageWrapper/PageWrapper';
@@ -16,20 +15,20 @@ import Register from '../Register/Register';
 import InfoToolTip from '../InfoTooltip/InfoTooltip';
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
 import LoadingError from '../LoadingError/LoadingError'
+import Navigation from "../Navigation/Navigation";
+
+import { CurrentUserContext } from '../../contexts/CurrentUserContext';
+
+import { useAuthRegister } from '../../utils/AuthRegHooks';
+import { useMoviesLogic } from '../../utils/FilteringLogicHooks';
 
 import Api from '../../utils/MainApi';
-import * as Auth from '../../utils/Auth';
-import { CurrentUserContext } from '../../contexts/CurrentUserContext';
-import Navigation from "../Navigation/Navigation";
-import { useMoviesLogic } from '../../utils/FilteringLogicHooks';
+
 
 
 function App(props) {
-    let navigate = useNavigate();
-
     const [isInfoTooltipOpen, setIsInfoTooltipOpen] = React.useState(false);
     const [isNavigationOpen, setIsNavigationOpen] = React.useState(false);
-    const [currentUser, setCurrentUser] = React.useState(null);
     const [tooltipMessage, setTooltipMessage] = React.useState('');
     const [tooltipIsOk, setTooltipIsOk] = React.useState(false);
     
@@ -39,33 +38,10 @@ function App(props) {
         setIsInfoTooltipOpen(true); 
     };
 
+    // открываем попап навигации по нажатию на бутер
     const handleNavigationOpen = ()=>setIsNavigationOpen(true);
-    
-    React.useEffect( 
-        ()=>{
-            const tkn = localStorage.getItem('token');
-            // проверяем токен пользователя
-            if (tkn) {
-                Auth.checkToken(tkn)
-                .then((res) => {
-                    if (res) { 
-                        //загружаем токен для запросов API
-                        Api.loadToken();
-                        //устанавливаем текущего пользователя
-                        setCurrentUser(res);
-                    }
-                })
-                .catch(err=>console.log(err))
-            }
-            //возвращаем функцию, которую вызовет реакт при завершении приложения
-            return(()=>{
-                localStorage.removeItem('searchParams');
-            });
-        },
-        [setCurrentUser] 
-    );
 
-    // подключаем логику работы с фильмами
+    // подключаем логику работы с фильмами из своего кастомного хука
     const {
         moviesList, 
         savedMoviesFiltered, 
@@ -78,74 +54,19 @@ function App(props) {
         clearAllMovies
     } = useMoviesLogic (Api);
 
+    // регистрация, авторизация
+    const {
+        currentUser,
+        handleRegister,
+        handleAuthorize,
+        handleSignOut,
+        handleUpdateUser
+    } = useAuthRegister(Api, handleInfoTolltipOpen, clearAllMovies)
+
+    // закрываем все попапы
     const closeAllPopups = () => {
         setIsInfoTooltipOpen(false);
         setIsNavigationOpen(false);
-    }
-
-    function handleRegister (name, email, password) {
-        Auth.register(name, email, password)
-        .then((res) => {
-            let msg;
-            let isAuthed=false;
-            if (res.code===200) {
-                console.log(res);
-                msg = 'Вы успешно \n зарегистрировались!';
-                isAuthed = true;
-            } else if (res.code===409) {
-                msg = res.body.message;
-            } else {
-                msg = 'Что-то пошло не так. \n Попробуйте еще раз'
-            }
-            handleInfoTolltipOpen(msg, isAuthed);
-        })
-        .catch(err=>console.log(err));
-    }
-
-    function handleAuthorize(email, password) {
-        Auth.authorize(email, password)
-        .then((res) => {
-            const {token, user} = res;
-            if (token){
-                localStorage.setItem('token',token);
-                // загружаем токен для запросов API
-                Api.loadToken();
-                // устанавливаем активного пользователя
-                setCurrentUser(user);
-                // переходим на страницу после логина
-                navigate('/');
-            } else {
-                handleInfoTolltipOpen('Неправильный \n логин или пароль', false);
-            }
-        })
-        .catch(err=>{
-            handleInfoTolltipOpen('Что-то пошло не так. \n попробуйте еще раз.', false);
-        });
-    }
-    
-    function handleSignOut() {
-        clearAllMovies();
-        localStorage.removeItem('token');
-        setCurrentUser(null);
-        navigate('/');
-    }
-
-    function handleUpdateUser(userName, email){
-        Api.patchProfile(email, userName)
-        .then ((res)=>{
-            setCurrentUser(res);
-            closeAllPopups();
-        })
-        .catch(err=>{
-            console.log(err)
-            if (err.status===409) {
-                handleInfoTolltipOpen('Указанный email уже занят. \n Попробуйте еще раз', false);
-            } else if (err.message==='Failed to fetch') {
-                handleInfoTolltipOpen('Проверьте соединение с интернетом. \n Попробуйте еще раз', false);
-            } else {
-                handleInfoTolltipOpen('Что-то пошло. \n Попробуйте еще раз', false);
-            }
-        })
     }
 
     return (
@@ -182,6 +103,7 @@ function App(props) {
                     } />
                     <Route path='/signin' element={<LogIn onAuthorise={handleAuthorize}/>} />
                     <Route path='/signup' element={<Register onRegister={handleRegister}/>} />
+                    <Route path='*' element={<LoadingError />} />
                 </Routes>
                 <Navigation opened={isNavigationOpen} onClose={closeAllPopups}/>
                 <InfoToolTip isOpen={isInfoTooltipOpen} tooltipMessage={tooltipMessage} tooltipIsOk={tooltipIsOk} onClose={closeAllPopups}/>
